We should be de-serializing an int64 for the UploadData identifier.
This is hard to test since this bug could only cause a very rare cache
collision.

R=eroman
BUG=25939
TEST=none

Review URL: http://codereview.chromium.org/338046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30231 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 69d320ef1ff94adc9ebfcb1230371aa0ae50a3e3

+++ chrome/common/common_param_traits.h
@@ -282,7 +282,7 @@ struct ParamTraits<scoped_refptr<net::UploadData> > {
     std::vector<net::UploadData::Element> elements;
     if (!ReadParam(m, iter, &elements))
       return false;
-    int identifier;
+    int64 identifier;
     if (!ReadParam(m, iter, &identifier))
       return false;
     *r = new net::UploadData;
